// src/types/projects.ts
export enum ProjectStatus {
  DONE = 'Done',
  IN_QUOTATION = 'In quotation',
  ACCEPTED = 'Accepted',
  NOT_STARTED = 'Not started',
  IN_PROCESS = 'In process',
  UNDER_REVISION = 'Under revision',
  DELAYED = 'Delayed',
  POSTPONED = 'Postponed',
  CANCELLED = 'Cancelled',
}

export enum DepartmentAreas {
  LEGAL = 'Legal',
  ACCOUNTING = 'Accounting',
  DIRECTIVES = 'Directives',
}